feat(modularize): separate out response builder
