+++ src/lib.rs
@@ -32,10 +32,6 @@ use futures::BoxFuture;
 
 use futures_cpupool::CpuPool;
 
-use serde_json::value::Value;
-// Rename type for crate
-type BTreeMap = std::collections::BTreeMap<String, Value>;
-
 mod request;
 use request::Request;
 
@@ -48,37 +44,8 @@ use error::Error;
 mod status;
 use status::Status;
 
-pub struct ResponseBuilder {
-    pub data: BTreeMap,
-    template: Option<String>,
-    status: Option<Status>,
-}
-
-impl ResponseBuilder {
-    pub fn new() -> ResponseBuilder {
-        ResponseBuilder {
-            data: BTreeMap::new(),
-            template: None,
-            status: None,
-        }
-    }
-
-    pub fn with_template(&mut self, template: String) {
-        self.template = Some(template);
-    }
-
-    pub fn with_status(&mut self, status: Status) {
-        self.status = Some(status);
-    }
-
-    pub fn to_response(self) -> Response {
-        Response {
-            data: self.data,
-            template: self.template.unwrap(),
-            status: self.status.unwrap(),
-        }
-    }
-}
+pub mod response_builder;
+pub use response_builder::ResponseBuilder;
 
 pub struct Server {
     routes: Vec<Route>,
